+++ src/main/java/com/JobTracker/demo/Service/SubContractorService.java
@@ -2,6 +2,7 @@ package com.JobTracker.demo.Service;
 
 import com.JobTracker.demo.Entity.SubContractor;
 import com.JobTracker.demo.Repository.SubContractorRepository;
+import com.JobTracker.demo.Repository.TaskRepository;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
@@ -11,9 +12,13 @@ import java.util.List;
 public class SubContractorService {
 
     private final SubContractorRepository subContractorRepository;
+    private final TaskRepository taskRepository;
 
-    public SubContractorService(SubContractorRepository subContractorRepository) {
+    public SubContractorService(SubContractorRepository subContractorRepository,
+                                TaskRepository taskRepository) {
         this.subContractorRepository = subContractorRepository;
+        this.taskRepository = taskRepository;
+
     }
 
     public List<SubContractor> findAll() {
@@ -33,18 +38,25 @@ public class SubContractorService {
         if(!subContractorRepository.existsById(subContractorId)) {
             throw new RuntimeException("sub contractor not found");
         }
+        if (taskRepository.existsBySubContractor_SubContractorId(subContractorId)) {
+            throw new IllegalArgumentException("Cannot delete sub-contractor: They are currently assigned to active tasks.");
+        }
         subContractorRepository.deleteById(subContractorId);
     }
 
     public SubContractor update(SubContractor subContractor, Long subContractorId) {
 
-        SubContractor currentSubContractor = subContractorRepository.findById(subContractorId)
-                .orElseThrow(() -> new RuntimeException("sub contractor not found"));
+        SubContractor currentSubContractor = findById(subContractorId);
 
-        currentSubContractor = subContractorRepository.findById(subContractorId).get();
-        currentSubContractor.setCompanyName(subContractor.getCompanyName());
-        currentSubContractor.setPrice(subContractor.getPrice());
-        currentSubContractor.setPhoneNumber(subContractor.getPhoneNumber());
+        if (subContractor.getCompanyName() != null && !subContractor.getCompanyName().isBlank()) {
+            currentSubContractor.setCompanyName(subContractor.getCompanyName());
+        }
+        if (subContractor.getPhoneNumber() != null && !subContractor.getPhoneNumber().isBlank()) {
+            currentSubContractor.setPhoneNumber(subContractor.getPhoneNumber());
+        }
+        if (subContractor.getPrice() != null) {
+            currentSubContractor.setPrice(subContractor.getPrice());
+        }
 
         return subContractorRepository.save(currentSubContractor);
     }
